rename haulerctl to just hauler
