+++ cmd/hauler/app/bundle_artifacts.go
