+++ cmd/hauler/app/relocate_artifacts.go
