+++ cmd/hauler/app/relocate_images.go
